feat: Allow passing home_id to save an API call to /me
In order to reduce calls to the Tado API, allow manually
passing of `home_id` to avoid a call to `/me` to retrieve it.

You can get `home_id` with the new method `t.get_home_id()` or
using the CLI via the existing `tado homes` command.

The `home_id` can also be passed as `-i` or `--home-id` to the CLI.